# Optimize sync partitions procedure