Summary
---

* Remove unnecessary instantiation of `FileStatus` for each directory being listed as a partition
* `system. sync_partition_metadata`: https://prestosql.io/docs/current/connector/hive.html#procedures

Performance Improvement
---

* running `system. sync_partition_metadata(FULL)`

| # Partitions  | Metastore | Time Before (s) | Time After (s) |
| ------------- | ------------- | ------------- | ------------- |
| 1000  | Glue | 90  | 4 |


**todo** more performance tests