@raunaqmorarka 2019-07-22
I think this change was already done with https://github.com/prestosql/presto/commit/bf7944cbc180182bee4f7c91f1b8c7479e80a85a