@bill-warshaw 2019-07-22
@raunaqmorarka I was based off the `316` tag locally because I couldn't get the build to work off of a snapshot version locally.

I'm confused about what exactly happened here, I see your commit in `master` and mine is not there.